+++ src/main/java/com/szh/monitor/service/OperationLogService.java
@@ -8,13 +8,13 @@ public interface OperationLogService {
 
     void logVisit(HttpServletRequest request);
 
-    void logCreate(String module, Integer targetId, String detail, HttpServletRequest request);
+    void logCreate(String module, Long targetId, String detail, HttpServletRequest request);
 
-    void logEdit(String module, Integer targetId, String detail, HttpServletRequest request);
+    void logEdit(String module, Long targetId, String detail, HttpServletRequest request);
 
-    void logDelete(String module, Integer targetId, String detail, HttpServletRequest request);
+    void logDelete(String module, Long targetId, String detail, HttpServletRequest request);
 
-    void log(String operationType, String module, Integer targetId, String detail, HttpServletRequest request);
+    void log(String operationType, String module, Long targetId, String detail, HttpServletRequest request);
 
     Page<OperationLog> getLogs(int page, int size);
 }
